+++ dev/src/reference.ts
@@ -16,13 +16,13 @@
 
 import * as firestore from '@google-cloud/firestore';
 import * as assert from 'assert';
+import {Duplex, Readable, Transform} from 'stream';
 import * as deepEqual from 'fast-deep-equal';
 import {GoogleError} from 'google-gax';
-import {Duplex, Readable, Transform} from 'stream';
 
 import * as protos from '../protos/firestore_v1_proto_api';
-import {Aggregate, AggregateField, AggregateSpec} from './aggregate';
 
+import {Aggregate, AggregateField, AggregateSpec} from './aggregate';
 import {
   DocumentSnapshot,
   DocumentSnapshotBuilder,
@@ -65,6 +65,7 @@ import {
 import {DocumentWatch, QueryWatch} from './watch';
 import {validateDocumentData, WriteBatch, WriteResult} from './write-batch';
 import api = protos.google.firestore.v1;
+import {ExplainMetrics, ExplainResults} from './query-profile';
 
 /**
  * The direction of a `Query.orderBy()` clause is specified as 'desc' or 'asc'
@@ -1685,7 +1686,9 @@ class QueryUtil<
     /** @private */
     readonly _firestore: Firestore,
     /** @private */
-    readonly _queryOptions: QueryOptions<AppModelType, DbModelType>
+    readonly _queryOptions: QueryOptions<AppModelType, DbModelType>,
+    /** @private */
+    readonly _serializer: Serializer
   ) {}
 
   _get(
@@ -1780,10 +1783,12 @@ class QueryUtil<
   _stream(
     query: Template,
     transactionIdOrReadTime?: Uint8Array | Timestamp,
-    retryWithCursor = true
+    retryWithCursor = true,
+    explainOptions?: firestore.ExplainOptions
   ): NodeJS.ReadableStream {
     const tag = requestTag();
     const startTime = Date.now();
+    const isExplain = explainOptions !== undefined;
 
     let lastReceivedDocument: QueryDocumentSnapshot<
       AppModelType,
@@ -1798,7 +1803,17 @@ class QueryUtil<
           callback(undefined);
           return;
         }
-        const readTime = Timestamp.fromProto(proto.readTime);
+
+        const output: {
+          readTime?: Timestamp;
+          document?: QueryDocumentSnapshot<AppModelType, DbModelType>;
+          explainMetrics?: ExplainMetrics;
+        } = {};
+
+        if (proto.readTime) {
+          output.readTime = Timestamp.fromProto(proto.readTime);
+        }
+
         if (proto.document) {
           const document = this._firestore.snapshot_(
             proto.document,
@@ -1818,16 +1833,24 @@ class QueryUtil<
             AppModelType,
             DbModelType
           >;
-          callback(undefined, {document: lastReceivedDocument, readTime});
-          if (proto.done) {
-            logger('Query._stream', tag, 'Trigger Logical Termination.');
-            backendStream.unpipe(stream);
-            backendStream.resume();
-            backendStream.end();
-            stream.end();
-          }
-        } else {
-          callback(undefined, {readTime});
+          output.document = lastReceivedDocument;
+        }
+
+        if (proto.explainMetrics) {
+          output.explainMetrics = ExplainMetrics._fromProto(
+            proto.explainMetrics,
+            this._serializer
+          );
+        }
+
+        callback(undefined, output);
+
+        if (proto.done) {
+          logger('QueryUtil._stream', tag, 'Trigger Logical Termination.');
+          backendStream.unpipe(stream);
+          backendStream.resume();
+          backendStream.end();
+          stream.end();
         }
       },
     });
@@ -1838,7 +1861,7 @@ class QueryUtil<
         // `_toProto()` might throw an exception. We rely on the behavior of an
         // async function to convert this exception into the rejected Promise we
         // catch below.
-        let request = query._toProto(transactionIdOrReadTime);
+        let request = query._toProto(transactionIdOrReadTime, explainOptions);
 
         let streamActive: Deferred<boolean>;
         do {
@@ -1855,12 +1878,15 @@ class QueryUtil<
 
             // If a non-transactional query failed, attempt to restart.
             // Transactional queries are retried via the transaction runner.
+            // Explain queries are not retried with a cursor. That would produce
+            // incorrect/partial profiling results.
             if (
+              !isExplain &&
               !transactionIdOrReadTime &&
               !this._isPermanentRpcError(err, 'runQuery')
             ) {
               logger(
-                'Query._stream',
+                'QueryUtil._stream',
                 tag,
                 'Query failed with retryable stream error:',
                 err
@@ -1873,7 +1899,7 @@ class QueryUtil<
               stream.write(NOOP_MESSAGE, () => {
                 if (this._hasRetryTimedOut(methodName, startTime)) {
                   logger(
-                    'Query._stream',
+                    'QueryUtil._stream',
                     tag,
                     'Query failed with retryable stream error but the total retry timeout has exceeded.'
                   );
@@ -1905,7 +1931,7 @@ class QueryUtil<
                   streamActive.resolve(/* active= */ true);
                 } else {
                   logger(
-                    'Query._stream',
+                    'QueryUtil._stream',
                     tag,
                     `Query failed with retryable stream error however either retryWithCursor="${retryWithCursor}", or ` +
                       'no progress was made receiving documents, so the stream is being closed.'
@@ -1916,7 +1942,7 @@ class QueryUtil<
               });
             } else {
               logger(
-                'Query._stream',
+                'QueryUtil._stream',
                 tag,
                 'Query failed with stream error:',
                 err
@@ -1995,7 +2021,7 @@ export class Query<
       AppModelType,
       DbModelType,
       Query<AppModelType, DbModelType>
-    >(_firestore, _queryOptions);
+    >(_firestore, _queryOptions, this._serializer);
   }
 
   /**
@@ -2949,6 +2975,91 @@ export class Query<
   }
 
   /**
+   * Plans and optionally executes this query. Returns a Promise that will be
+   * resolved with the planner information, statistics from the query execution (if any),
+   * and the query results (if any).
+   *
+   * @return A Promise that will be resolved with the planner information, statistics
+   *  from the query execution (if any), and the query results (if any).
+   */
+  explain(
+    options?: firestore.ExplainOptions
+  ): Promise<ExplainResults<QuerySnapshot<AppModelType, DbModelType>>> {
+    if (options === undefined) {
+      options = {};
+    }
+
+    // Capture the error stack to preserve stack tracing across async calls.
+    const stack = Error().stack!;
+
+    return new Promise((resolve, reject) => {
+      let readTime: Timestamp;
+      let docs: Array<QueryDocumentSnapshot<AppModelType, DbModelType>> | null =
+        null;
+      let metrics: ExplainMetrics | null = null;
+
+      this._stream(undefined, options)
+        .on('error', err => {
+          reject(wrapError(err, stack));
+        })
+        .on('data', data => {
+          if (data.readTime) {
+            readTime = data.readTime;
+          }
+          if (data.document) {
+            if (docs === null) {
+              docs = [];
+            }
+            docs.push(data.document);
+          }
+          if (data.explainMetrics) {
+            metrics = data.explainMetrics;
+
+            if (docs === null && metrics?.executionStats !== null) {
+              // This indicates that the query was executed, but no documents
+              // had matched the query.
+              docs = [];
+            }
+          }
+        })
+        .on('end', () => {
+          if (metrics === null) {
+            reject('No explain results.');
+          }
+
+          // Some explain queries will not have a snapshot associated with them.
+          let snapshot: QuerySnapshot<AppModelType, DbModelType> | null = null;
+          if (docs !== null) {
+            if (this._queryOptions.limitType === LimitType.Last) {
+              // The results for limitToLast queries need to be flipped since
+              // we reversed the ordering constraints before sending the query
+              // to the backend.
+              docs.reverse();
+            }
+
+            snapshot = new QuerySnapshot(
+              this,
+              readTime,
+              docs.length,
+              () => docs!,
+              () => {
+                const changes: Array<
+                  DocumentChange<AppModelType, DbModelType>
+                > = [];
+                for (let i = 0; i < docs!.length; ++i) {
+                  changes.push(new DocumentChange('added', docs![i], -1, i));
+                }
+                return changes;
+              }
+            );
+          }
+
+          resolve(new ExplainResults(metrics!, snapshot));
+        });
+    });
+  }
+
+  /**
    * Internal get() method that accepts an optional transaction id.
    *
    * @private
@@ -2990,6 +3101,62 @@ export class Query<
   }
 
   /**
+   * Executes the query and streams the results as the following object:
+   * {document?: DocumentSnapshot, metrics?: ExplainMetrics}
+   *
+   * The stream surfaces documents one at a time as they are received from the
+   * server, and at the end, it will surface the metrics associated with
+   * executing the query.
+   *
+   * @example
+   * ```
+   * let query = firestore.collection('col').where('foo', '==', 'bar');
+   *
+   * let count = 0;
+   *
+   * query.explainStream({analyze: true}).on('data', (data) => {
+   *   if (data.document) {
+   *     // Use data.document which is a DocumentSnapshot instance.
+   *     console.log(`Found document with name '${data.document.id}'`);
+   *     ++count;
+   *   }
+   *   if (data.metrics) {
+   *     // Use data.metrics which is an ExplainMetrics instance.
+   *   }
+   * }).on('end', () => {
+   *   console.log(`Received ${count} documents.`);
+   * });
+   * ```
+   */
+  explainStream(
+    explainOptions?: firestore.ExplainOptions
+  ): NodeJS.ReadableStream {
+    if (explainOptions === undefined) {
+      explainOptions = {};
+    }
+    if (this._queryOptions.limitType === LimitType.Last) {
+      throw new Error(
+        'Query results for queries that include limitToLast() ' +
+          'constraints cannot be streamed. Use Query.explain() instead.'
+      );
+    }
+
+    const responseStream = this._stream(undefined, explainOptions);
+    const transform = new Transform({
+      objectMode: true,
+      transform(chunk, encoding, callback) {
+        callback(undefined, {
+          document: chunk.document,
+          metrics: chunk.explainMetrics,
+        });
+      },
+    });
+    responseStream.pipe(transform);
+    responseStream.on('error', e => transform.destroy(e));
+    return transform;
+  }
+
+  /**
    * Converts a QueryCursor to its proto representation.
    *
    * @param cursor The original cursor value
@@ -3012,12 +3179,14 @@ export class Query<
    *
    * @param transactionIdOrReadTime A transaction ID or the read time at which
    * to execute the query.
+   * @param explainOptions Options to use for explaining the query (if any).
    * @private
    * @internal
    * @returns Serialized JSON for the query.
    */
   _toProto(
-    transactionIdOrReadTime?: Uint8Array | Timestamp
+    transactionIdOrReadTime?: Uint8Array | Timestamp,
+    explainOptions?: firestore.ExplainOptions
   ): api.IRunQueryRequest {
     const projectId = this.firestore.projectId;
     const databaseId = this.firestore.databaseId;
@@ -3069,6 +3238,10 @@ export class Query<
     } else if (transactionIdOrReadTime instanceof Timestamp) {
       runQueryRequest.readTime =
         transactionIdOrReadTime.toProto().timestampValue;
+    }
+
+    if (explainOptions) {
+      runQueryRequest.explainOptions = explainOptions;
     }
 
     return runQueryRequest;
@@ -3148,14 +3321,21 @@ export class Query<
    *
    * @param transactionIdOrReadTime A transaction ID or the read time at which
    * to execute the query.
+   * @param explainOptions Options to use for explaining the query (if any).
    * @private
    * @internal
    * @returns A stream of document results.
    */
   _stream(
-    transactionIdOrReadTime?: Uint8Array | Timestamp
+    transactionIdOrReadTime?: Uint8Array | Timestamp,
+    explainOptions?: firestore.ExplainOptions
   ): NodeJS.ReadableStream {
-    return this._queryUtil._stream(this, transactionIdOrReadTime);
+    return this._queryUtil._stream(
+      this,
+      transactionIdOrReadTime,
+      true,
+      explainOptions
+    );
   }
 
   /**
@@ -3750,17 +3930,28 @@ export class AggregateQuery<
     // Capture the error stack to preserve stack tracing across async calls.
     const stack = Error().stack!;
 
+    let result: AggregateQuerySnapshot<
+      AggregateSpecType,
+      AppModelType,
+      DbModelType
+    > | null = null;
+
     return new Promise((resolve, reject) => {
       const stream = this._stream(transactionIdOrReadTime);
       stream.on('error', err => {
         reject(wrapError(err, stack));
       });
-      stream.once('data', result => {
-        stream.destroy();
-        resolve(result);
+      stream.on('data', data => {
+        if (data.aggregationResult) {
+          result = data.aggregationResult;
+        }
       });
       stream.on('end', () => {
-        reject('No AggregateQuery results');
+        stream.destroy();
+        if (result === null) {
+          reject(Error('RunAggregationQueryResponse is missing result'));
+        }
+        resolve(result!);
       });
     });
   }
@@ -3770,23 +3961,48 @@ export class AggregateQuery<
    *
    * @private
    * @internal
-   * @param transactionId A transaction ID.
+   * @param transactionIdOrReadTime A transaction ID or the read time at which
+   * to execute the query.
+   * @param explainOptions Options to use for explaining the query (if any).
    * @returns A stream of document results.
    */
-  _stream(transactionIdOrReadTime?: Uint8Array | Timestamp): Readable {
+  _stream(
+    transactionIdOrReadTime?: Uint8Array | Timestamp,
+    explainOptions?: firestore.ExplainOptions
+  ): Readable {
     const tag = requestTag();
     const firestore = this._query.firestore;
 
     const stream: Transform = new Transform({
       objectMode: true,
       transform: (proto: api.IRunAggregationQueryResponse, enc, callback) => {
+        const output: {
+          aggregationResult?: AggregateQuerySnapshot<
+            AggregateSpecType,
+            AppModelType,
+            DbModelType
+          >;
+          explainMetrics?: ExplainMetrics;
+        } = {};
+
         if (proto.result) {
           const readTime = Timestamp.fromProto(proto.readTime!);
           const data = this.decodeResult(proto.result);
-          callback(undefined, new AggregateQuerySnapshot(this, readTime, data));
-        } else {
-          callback(Error('RunAggregationQueryResponse is missing result'));
+          output.aggregationResult = new AggregateQuerySnapshot(
+            this,
+            readTime,
+            data
+          );
         }
+
+        if (proto.explainMetrics) {
+          output.explainMetrics = ExplainMetrics._fromProto(
+            proto.explainMetrics,
+            firestore._serializer!
+          );
+        }
+
+        callback(undefined, output);
       },
     });
 
@@ -3796,7 +4012,7 @@ export class AggregateQuery<
         // `_toProto()` might throw an exception. We rely on the behavior of an
         // async function to convert this exception into the rejected Promise we
         // catch below.
-        const request = this.toProto(transactionIdOrReadTime);
+        const request = this.toProto(transactionIdOrReadTime, explainOptions);
 
         const backendStream = await firestore.requestStream(
           'runAggregationQuery',
@@ -3813,6 +4029,7 @@ export class AggregateQuery<
           // consider implementing retries if the stream is making progress
           // receiving results for groups. See the use of lastReceivedDocument
           // in the retry strategy for runQuery.
+          // Also note that explain queries should not be retried.
 
           backendStream.unpipe(stream);
           logger(
@@ -3869,7 +4086,8 @@ export class AggregateQuery<
    * @returns Serialized JSON for the query.
    */
   toProto(
-    transactionIdOrReadTime?: Uint8Array | Timestamp
+    transactionIdOrReadTime?: Uint8Array | Timestamp,
+    explainOptions?: firestore.ExplainOptions
   ): api.IRunAggregationQueryRequest {
     const queryProto = this._query._toProto();
     const runQueryRequest: api.IRunAggregationQueryRequest = {
@@ -3895,6 +4113,10 @@ export class AggregateQuery<
       runQueryRequest.transaction = transactionIdOrReadTime;
     } else if (transactionIdOrReadTime instanceof Timestamp) {
       runQueryRequest.readTime = transactionIdOrReadTime;
+    }
+
+    if (explainOptions) {
+      runQueryRequest.explainOptions = explainOptions;
     }
 
     return runQueryRequest;
@@ -3929,6 +4151,62 @@ export class AggregateQuery<
       return false;
     }
     return deepEqual(this._aggregates, other._aggregates);
+  }
+
+  /**
+   * Plans and optionally executes this query. Returns a Promise that will be
+   * resolved with the planner information, statistics from the query
+   * execution (if any), and the query results (if any).
+   *
+   * @return A Promise that will be resolved with the planner information,
+   * statistics from the query execution (if any), and the query results (if any).
+   */
+  explain(
+    options?: firestore.ExplainOptions
+  ): Promise<
+    ExplainResults<
+      AggregateQuerySnapshot<AggregateSpecType, AppModelType, DbModelType>
+    >
+  > {
+    if (options === undefined) {
+      options = {};
+    }
+    // Capture the error stack to preserve stack tracing across async calls.
+    const stack = Error().stack!;
+
+    let metrics: ExplainMetrics | null = null;
+    let aggregationResult: AggregateQuerySnapshot<
+      AggregateSpecType,
+      AppModelType,
+      DbModelType
+    > | null = null;
+
+    return new Promise((resolve, reject) => {
+      const stream = this._stream(undefined, options);
+      stream.on('error', err => {
+        reject(wrapError(err, stack));
+      });
+      stream.on('data', data => {
+        if (data.aggregationResult) {
+          aggregationResult = data.aggregationResult;
+        }
+
+        if (data.explainMetrics) {
+          metrics = data.explainMetrics;
+        }
+      });
+      stream.on('end', () => {
+        stream.destroy();
+        if (metrics === null) {
+          reject('No explain results.');
+        }
+        resolve(
+          new ExplainResults<
+            AggregateQuerySnapshot<AggregateSpecType, AppModelType, DbModelType>
+          >(metrics!, aggregationResult)
+        );
+      });
+    });
   }
 }
 
@@ -4078,7 +4356,7 @@ export class VectorQuery<
       AppModelType,
       DbModelType,
       VectorQuery<AppModelType, DbModelType>
-    >(_query._firestore, _query._queryOptions);
+    >(_query._firestore, _query._queryOptions, _query._serializer);
   }
 
   /** The query whose results participants in the vector search. Filtering